+++ src/hooks/index.ts
@@ -1,8 +1,9 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { auth, create, deleteOne, getList, getOne, update } from "../providers";
-import { message } from "antd";
+import { message, notification } from "antd";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
+import { jwtDecode } from "jwt-decode";
 
 type Props = {
   resource: string;
@@ -58,33 +59,55 @@ export const useDelete = ({ resource = "products" }: Props) => {
   });
 };
 
-export const useAuth = ({ resource = "register" }) => {
+export const useAuth = ({ resource = "register" }: Props) => {
   const nav = useNavigate();
-
   return useMutation({
-    mutationFn: async (values: any) => {
-      const { repassword, ...userData } = {
-        ...values,
-        role: "user",
-        status: "active",
-      };
-
-      return auth({resource, values:userData}).then((res) => res.data);
-    },
+    mutationFn: (values: any) => auth({ resource, values }),
     onSuccess: (data) => {
-      message.success("Thành công!");
+      message.success("thanh cong");
+      if (resource == "register") {
+        nav("/login");
+        return;
+      }
+     
 
-      if (resource === "register") {
-        nav("/login"); // Nếu đăng ký, chuyển hướng đến trang đăng nhập
+      const { accessToken, user } = data;
+      const decodedToken: any = jwtDecode(accessToken);
+
+      console.log(decodedToken);
+
+      if (user.status === "banned") {
+        notification.error({
+          message: "Login Failed",
+          description: "Tài khoản của bạn đã bị khóa.",
+          duration: 5,
+        });
         return;
       }
 
-      localStorage.setItem("token", data.accessToken);
-      localStorage.setItem("user", JSON.stringify(data.user));
-      
+      const userData = {
+        id: user.id,
+        name: user.name,
+        email: user.email,
+        phone: user.phone,
+      };
+
+      localStorage.setItem("token", accessToken);
+      localStorage.setItem("role", user.role);
+      localStorage.setItem("user", JSON.stringify(userData));
+
+      notification.success({ message: "Đăng nhập thành công!" });
+
+      if (user.role === "user") {
+        nav("/");
+        window.location.reload();
+      } else {
+        nav("/admin");
+      }
     },
-    onError: (error) => {
-      message.error("Thất bại: "+ error.message);
-    },
+
+    
+    onError: () => {},
   });
 };
+
